reformed trivia question function
Moved initial and edited question timer to ephemeral message.  New message with outcome is sent to the channel as a separate message from the interactions.  Original message is edited and able to be cleared by the user.   Added points + bonus to trivia question embed,  and split them in the response embed for correct answer to show how many points + bonus earned.   Added command user's mention in the embed and their avatar as embed URL, if exists.