more minor changes
Switched the original message away from ephemeral. Decided it would be better, at least for my guild members to just have that original message auto delete after the results message is sent.  Added timestamp to the embed., and if the member has an avatar, it's also included in the embeds.   Made an small adjustment to the view timeout limit since it was sometimes ended while still showing 3+ seconds left on the trivia question embed.